requirements.txt
# Flask framework and WSGI server
Flask==3.0.3
gunicorn==23.0.0

# Firebase Admin SDK for Firestore
firebase-admin==6.5.0

# Google Cloud Firestore
google-cloud-firestore==2.19.0

# Requests library for API calls
requests==2.32.3

# Slack Bolt for Slack bot
slack-bolt==1.20.1
slack-sdk==3.33.1

# dotenv for loading environment variables
python-dotenv==1.0.1

# Regex for validations
regex==2024.9.11

asyncio==3.4.3

Flask-Script==2.0.6

slackeventsapi==3.0.3